Further attempt to fix number display in windows

+++ src/cmd-pickup.c
@@ -92,8 +92,7 @@ static void player_pickup_gold(void)
 
 		/* Build a message */
 		(void)strnfmt(buf, sizeof(buf),
-					  "You have found %ld gold pieces worth of ",
-					  (long)total_gold);
+					  "You have found %d gold pieces worth of ", total_gold);
 
 		/* One treasure type.. */
 		if (at_most_one)
